STUD-74381 Can't select global variables in activity designer (#352)
* STUD-74381 Can't select global variables in activity designer

In JitCompiler we need to resolve assemblies that may have been loaded in a different AssemblyLoadContext. We do this on the Resolving event.

In the past we also called loadContext.LoadFromAssemblyName(assemblyName) which entered a loop throwing StackOverflowException. This call is not needed since any missing assembly will trigger an error earlier in the compilation step.