STUD-76393 Packager: Slow compilation for large projects (#367)
Add static global cache for ExternalMetadataReferenceResolver. This greatly improves performance and memory consumption. When this resolver is used in a compilation, ResolveMissingAssembly is called very often and in projects with a lot of files (can be thousands) we created metadata reference for the same assembly multiple times for the same file.

https://uipath.atlassian.net/browse/STUD-76393